backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
fastapi==0.115.0
fonttools==4.61.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.9.2
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.0
preshed==3.0.12
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
pyparsing==3.3.2
pyspellchecker==0.8.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.2
starlette==0.38.6
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.2
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==16.0
wrapt==2.1.1
bcrypt==4.1.3
PyJWT==2.8.0
psycopg2-binary==2.9.9